requirements.txt
requests
pysocks
goose3
lxml
python-dateutil
Faker
pyppeteer